feat: flash package instead of rootfs, major refactor
