+++ cli/Cargo.toml
@@ -6,6 +6,9 @@ edition = "2024"
 [dependencies]
 clap = { version = "4.5.32", features = ["derive"] }
 colored = "3.0.0"
-include_dir = "0.7.4"
 inquire = "0.7.5"
+serde = { version = "1.0.219", features = ["derive"] }
+serde_yml = "0.0.12"
+tempfile = "3.19.1"
 uuu-rs = { path = "../lib/uuu-rs" }
+zip = "2.5.0"
