+++ cli/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod cli;
 pub mod script;
+pub mod commands;
